refactor tables fetch
